Fixed - Map Volume: Resetting itself because of ACE Hearing
What
=================
Fixed: Map lower volume was breaking whenever ACEX was not being loaded,
because ACEX would disable the ACE Hearing feature temporarily everytime
it lowered its volume. ACE Hearing would update the sound volume
depending on how "deaf" the player is. This of course, because it was
running every frame, would override what we were trying to do in
lowering the volume whilst in a vehicle and on the map.

Issue ID: N/A