Added/Removed - Arsenal stats for radio equipment & obsolete SATCOM fix
What
=================
Added: ACE Arsenal stats "Extended Range" and "Range Boost Factor" as
niche stats for when the Whip antenna is deployed on a backpack, and the
boost factor that the SATCOM antenna has as an effect on the radio
range of the user.
Removed: An obsolete fix that was to make the SATCOM ACE compatible. The
ILBE - Rework mod had already fixed this for a long time, we just didn't
notice until now.

Issue ID: N/A