Fixed - Respawn Timer resetting after each death
What
=================
Since Arma 3 update v2.00 it seems the respawn timer needs to be set
each time a person dies, otherwise it would ignore the modified respawn
timers set after the next death and respawn was completed, thus
resulting in a timer reset to the same time that the mission has set
from the beginning of the mission.
This made much more sense in my head... anyway, fixed!

Issue ID: #3